backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
google-generativeai==0.3.1
chromadb==0.4.15
sentence-transformers==2.7.0
huggingface-hub==0.20.0
PyPDF2==3.0.1
python-docx==1.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
